+++ tools/update_version/update_version.py
@@ -14,39 +14,6 @@ import sys
 import json
 
 
-def get_appveyor_repo_commit():
-    return os.getenv('APPVEYOR_REPO_COMMIT')
-
-
-def get_appveyor_build_number():
-    return int(os.getenv('APPVEYOR_BUILD_NUMBER'))
-
-
-def get_appveyor_build_version():
-    return os.getenv('APPVEYOR_BUILD_VERSION')
-
-
-def use_appveyor_variables():
-    return get_appveyor_repo_commit() is not None and get_appveyor_build_number() is not None and get_appveyor_build_version() is not None
-
-
-def edit_appveyor_yml(major, minor, maintenance):
-    lines_out = []
-    filename = './appveyor.yml'
-    with open(filename) as f:
-        lines_in = f.readlines()
-        for line in lines_in:
-            line = line.replace('\r\n', '')
-            line = line.replace('\n', '')
-            if line.strip().startswith('version:'):
-                lines_out.append('version: ' + str(major) + '.' +
-                                 str(minor) + '.' + str(maintenance) + '.{build}')
-            else:
-                lines_out.append(line)
-    with open(filename, 'w') as f:
-        for l in lines_out:
-            f.write(l + '\n')
-
 
 def is_dirty_git():
     status = subprocess.check_output(['git', 'status', '-uno', '--porcelain'])
@@ -259,17 +226,7 @@ if __name__ == '__main__':
     current_version = get_current_version()
     update_from_command_line = False
 
-    if use_appveyor_variables() is True:
-        print('USE APPVEYOR')
-        print('REPO COMMIT: ' + get_appveyor_repo_commit())
-        print('BUILD NUMBER: ' + str(get_appveyor_build_number()))
-        print('BUILD VERSION: ' + get_appveyor_build_version())
-        major = current_version[0]
-        minor = current_version[1]
-        maintenance = current_version[2]
-        build = get_appveyor_build_number()
-        git_hash = get_appveyor_repo_commit()
-    elif use_github_variables() is True:
+    if use_github_variables() is True:
         print('USE GITHUB')
         print('REPO COMMIT: ' + get_github_repo_commit())
         print('BUILD NUMBER: ' + str(get_github_build_number()))
@@ -319,7 +276,6 @@ if __name__ == '__main__':
     edit_nelson_version_h_vc(major, minor, maintenance, build, git_hash)
     edit_nelson_version_h_in(git_hash)
     if update_from_command_line == True:
-        edit_appveyor_yml(major, minor, maintenance)
         edit_package_json(major, minor, maintenance)
     edit_homepage_md(version_str)
     sys.exit(0)