+++ CHANGELOG.md
@@ -27,6 +27,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - [#702](http://github.com/Nelson-numerical-software/nelson/issues/702): Ubuntu 18.04 CI removed on Github.
 
+- Remove definitively Appveyor CI.
+
 ### Fixed
 
 - [#723](http://github.com/Nelson-numerical-software/nelson/issues/723): MacOS CI was broken.